@@ -2,7 +2,7 @@ name: starguide_client
 description: Starting point for a Serverpod client.
 
 environment:
-  sdk: ">=3.3.0 <4.0.0"
+  sdk: ^3.8.0
 
 dependencies:
   serverpod_client: ^2.8.0
