@@ -2,7 +2,7 @@ name: starguide_server
 description: Starting point for a Serverpod server.
 
 environment:
-  sdk: ">=3.3.0 <4.0.0"
+  sdk: ^3.8.0
 
 dependencies:
   http: ^1.3.0
