@@ -2,15 +2,17 @@ import 'package:starguide_server/src/business/data_fetcher.dart';
 import 'package:starguide_server/src/business/data_sources/github_docs.dart';
 
 DataFetcher setupDataFetcher() {
-  return DataFetcher(dataSources: [
-    GithubDocsDataSource(
-      owner: 'serverpod',
-      repo: 'serverpod_docs',
-      basePath: 'docs',
-      referenceUrl: Uri.parse('https://docs.serverpod.dev'),
-      branch: 'main',
-    ),
-  ]);
+  return DataFetcher(
+    dataSources: [
+      GithubDocsDataSource(
+        owner: 'serverpod',
+        repo: 'serverpod_docs',
+        basePath: 'docs',
+        referenceUrl: Uri.parse('https://docs.serverpod.dev'),
+        branch: 'main',
+      ),
+    ],
+  );
 }
 
 // TODO: Github Discussions
