@@ -2,8 +2,7 @@ import 'package:test/test.dart';
 import 'package:starguide_server/src/example_util.dart';
 
 void main() {
-  test(
-      'Given ExampleUtil '
+  test('Given ExampleUtil '
       'when calling `buildGreeting` with name '
       'then returned greeting includes name', () async {
     final greeting = ExampleUtil.buildGreeting('Alice');
